refactoring to put common functions in ctrlplot.py
