+++ control/freqplot.py
@@ -19,14 +19,14 @@ import numpy as np
 
 from . import config
 from .bdalg import feedback
-from .ctrlplot import suptitle, _find_axes_center
+from .ctrlplot import suptitle, _find_axes_center, _make_legend_labels, \
+    _update_suptitle
 from .ctrlutil import unwrap
 from .exception import ControlMIMONotImplemented
 from .frdata import FrequencyResponseData
 from .lti import LTI, _process_frequency_response, frequency_response
 from .margins import stability_margins
 from .statesp import StateSpace
-from .timeplot import _make_legend_labels
 from .xferfcn import TransferFunction
 
 __all__ = ['bode_plot', 'NyquistResponseData', 'nyquist_response',
@@ -954,28 +954,7 @@ def bode_plot(
         else:
             title = data[0].title
 
-    if fig is not None and isinstance(title, str):
-        # Get the current title, if it exists
-        old_title = None if fig._suptitle is None else fig._suptitle._text
-        new_title = title
-
-        if old_title is not None:
-            # Find the common part of the titles
-            common_prefix = commonprefix([old_title, new_title])
-
-            # Back up to the last space
-            last_space = common_prefix.rfind(' ')
-            if last_space > 0:
-                common_prefix = common_prefix[:last_space]
-            common_len = len(common_prefix)
-
-            # Add the new part of the title (usually the system name)
-            if old_title[common_len:] != new_title[common_len:]:
-                separator = ',' if len(common_prefix) > 0 else ';'
-                new_title = old_title + separator + new_title[common_len:]
-
-        # Add the title
-        suptitle(title, fig=fig, rcParams=rcParams, frame=suptitle_frame)
+    _update_suptitle(fig, title, rcParams=rcParams, frame=suptitle_frame)
 
     #
     # Create legends
